Refactor file operations and commands tool for improved structure and functionality
